Add files via upload
Many traffic light systems operate on a timing mechanism that changes the lights after a given interval. A Smart traffic light is a controller that senses the presence or absence of vehicles and reacts accordingly. The idea behind Smart traffic light controllers is that drivers will not spend unnecessary time waiting for the traffic lights to change. A Smart traffic Light-Controller detects traffic by using Sensors. 
   Auto Traffic Detection
• Correct Scheduled timing
• Reduces the waiting time